refactored data.py in data/ , no code changes
